Fix 'invalid thread id' issue #13 by making sure tid is a C long int.

+++ src/pyshark/utils.py
@@ -22,7 +22,7 @@ def _async_raise(tid, exctype):
     """
     if not inspect.isclass(exctype):
         raise TypeError("Only types can be raised (not instances)")
-    res = ctypes.pythonapi.PyThreadState_SetAsyncExc(tid,
+    res = ctypes.pythonapi.PyThreadState_SetAsyncExc(ctypes.c_long(tid),
                                                   ctypes.py_object(exctype))
     if res == 0:
         raise ValueError("invalid thread id")
